make substance container nested class
